requirements.txt
stravaio
notion-client
python-dotenv